ci: update CI conf to work with the new MacOSX sdk
To support Apple ARM arch we need macOS sdk vers >= 11.0.
We chose to adopt the latest stable version for the 11.x
branch, which is 11.3.

Other than the the new GNU tools that let us detect which
kiond of arch we are building for change the way to label
apple architectures, namely:

- `x86_64-apple-darwin` for Intel macOS
- `arm64-apple-darwin` for ARM macOS

In a subsequent commit we will add a new CI task to build
Apple ARM compatible NEXA binaries.