Replace $(AT) with .SILENCE.
This reduces the amount of syntax noise in the makefiles.
Setting V=1 still enables verbose logging.

The only noticeable difference in behavior is that, unless V=1 is specified, make won't print its own messages like
make: Nothing to be done for 'all', make: 'all' is up to date, or touch <file>, if -t is specified.