Update deprecated to fix compile warnings (qt)
Fixes many of the compile warnings (from deprecated declarations) when compiling with current QT package from depends.

We still have quite a bit of warnings on more recent versions of Qt (>5.12), but we can't do much about it unless we decide to drop support of Ubuntu 20.04, which is not something we would do soon.