Relax Berkeley DB version restriction when using `--with-incompatible-bdb` configure flag
Running `./configure` w/o the incpat flag would succeed only if berkeley db 5.3 is installed.

Running `./configure --with-incompatible-bdb` would succeed only if berkeley db 4.8 or 18.1 are installed (on linux only 4.8 is available via distro packages).

In case only 4.8 or 18.1 are installed then the only way to get configure not failing would be to use `--with-incompatible-bdb` flag. In case both 4.3 and 18.1 are installed, 18.1 will be chosen.