Fix building on FreeBSD and update the documentation
Mostly I just had to add FreeBSD to the list of supported operating
systems in various places. In one place I had to change the path to find
the path of the executable because it's different in FreeBSD.

Unfortunately in FreeBSD the only easy way (see
https://stackoverflow.com/a/1024937/428740) to get this information
programtically requires configuring procfs which I have documented. I
searched the code and it's only used in one place if you use rostrum.